Refactor employee turnover prediction model to use simple logistic regression; update requirements and add setup script for Docker environment

+++ dbt/models/intermediate/int_engagement_metrics.sql
@@ -21,7 +21,7 @@ engagement_metrics as (
         engagement_score,
         satisfaction_score,
         work_life_balance_score,
-        overall_wellbeing_score,
+        overall_wellbeing_score::float,
         engagement_level,
         satisfaction_level,
         work_life_balance_level,
